初版

+++ src/main/java/com/top/logisticsStage/service/T_MANUAL_NEWS_EC_LIST_STATEService.java
@@ -7,6 +7,7 @@ import com.top.logisticsStage.repository.T_MANUAL_EST_ECRepository;
 import com.top.logisticsStage.repository.T_MANUAL_NEWS_EC_LIST_STATERepository;
 import com.top.logisticsStage.service.dto.T_MANUAL_NEWS_EC_LIST_STATEDTO;
 import com.top.logisticsStage.service.mapper.T_MANUAL_NEWS_EC_LIST_STATEMapper;
+import com.top.logisticsStage.web.rest.vm.T_MANUAL_EST_ECQueryVM;
 import com.top.logisticsStage.web.rest.vm.T_MANUAL_NEWS_EC_LIST_STATEQueryVM;
 import org.apache.commons.lang.StringUtils;
 import org.hibernate.service.spi.ServiceException;
@@ -25,6 +26,7 @@ import javax.persistence.criteria.Root;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -59,10 +61,13 @@ public class T_MANUAL_NEWS_EC_LIST_STATEService {
         log.info(SecurityUtils.getCurrentUserLogin() +"修改新数据："+dto.toString());
         T_MANUAL_NEWS_EC_LIST_STATE entity = t_MANUAL_NEWS_EC_LIST_STATEMapper.toEntity(dto);
         entity = t_MANUAL_NEWS_EC_LIST_STATERepository.save(entity);
-        t_MANUAL_NEWS_EC_LIST_STATERepository.deleteById(dto.getOldItemCode());
-        List<T_MANUAL_EST_EC> list1 = t_MANUAL_EST_ECRepository.findAllByItemCode(dto.getOldItemCode());
-        list1.forEach(e->e.setItemCode(dto.getItemCode()));
-        t_MANUAL_EST_ECRepository.saveAll(list1);
+        //关联新品目标表同步更新
+        if(!dto.getOldItemCode().equals(dto.getItemCode())) {
+            t_MANUAL_NEWS_EC_LIST_STATERepository.deleteById(dto.getOldItemCode());
+            List<T_MANUAL_EST_EC> list1 = t_MANUAL_EST_ECRepository.findAllByItemCode(dto.getOldItemCode());
+            list1.forEach(e -> e.setItemCode(dto.getItemCode()));
+            t_MANUAL_EST_ECRepository.saveAll(list1);
+        }
         return t_MANUAL_NEWS_EC_LIST_STATEMapper.toDto(entity);
     }
 
@@ -70,6 +75,13 @@ public class T_MANUAL_NEWS_EC_LIST_STATEService {
         Page<T_MANUAL_NEWS_EC_LIST_STATE> list = t_MANUAL_NEWS_EC_LIST_STATERepository.findAll(buildMultConditional(queryVM), pageable);
         Page<T_MANUAL_NEWS_EC_LIST_STATEDTO> page = list.map(t_MANUAL_NEWS_EC_LIST_STATEMapper::toDto);
         return page;
+    }
+
+    public Integer deleteByVm(T_MANUAL_NEWS_EC_LIST_STATEQueryVM queryVM) {
+        List<T_MANUAL_NEWS_EC_LIST_STATE> list =t_MANUAL_NEWS_EC_LIST_STATERepository.findAll(buildMultConditional(queryVM));
+        List<String> ids = list.stream().map(T_MANUAL_NEWS_EC_LIST_STATE::getItemCode).collect(Collectors.toList());
+        Integer flg = t_MANUAL_NEWS_EC_LIST_STATERepository.deleteByItemCodeIn(ids);
+        return flg;
     }
 
     private Specification<T_MANUAL_NEWS_EC_LIST_STATE> buildMultConditional(T_MANUAL_NEWS_EC_LIST_STATEQueryVM queryVM) {
@@ -87,6 +99,18 @@ public class T_MANUAL_NEWS_EC_LIST_STATEService {
                 }
                 if (StringUtils.isNotBlank(queryVM.getOnTrace())) {
                     Predicate p= cb.equal(root.<String>get("onTrace"), queryVM.getOnTrace());
+                    predicates.add(p);
+                }
+                if (StringUtils.isNotBlank(queryVM.getCharge())) {
+                    Predicate p= cb.like(root.<String>get("charge"), "%" + queryVM.getCharge() + "%");
+                    predicates.add(p);
+                }
+                if (StringUtils.isNotBlank(queryVM.getSeries())) {
+                    Predicate p= cb.like(root.<String>get("series"), "%" + queryVM.getSeries() + "%");
+                    predicates.add(p);
+                }
+                if (StringUtils.isNotBlank(queryVM.getDivision())) {
+                    Predicate p= cb.like(root.<String>get("division"), "%" + queryVM.getDivision() + "%");
                     predicates.add(p);
                 }
                 if (!predicates.isEmpty()) {
